+++ callback_model/algo2.hpp
@@ -3,10 +3,10 @@
 
 class Algo2 : public IMarketDataHandler {
 public:
-    Algo2(int instrument);
+    Algo2(long long int instrument);
     void OnPriceUpdate(PriceUpdate priceUpdate) override final;
     void start(ng_sdk& sdk_api);
 
 private:
-    int _instrument;
+    long long int _instrument;
 };
